requirements.txt
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
tqdm==4.65.0
python-dateutil==2.8.2
transformers>=4.30.0
torch>=2.0.0
tqdm>=4.65.0
accelerate>=0.20.0